fix(java): decode scoped-connection responses larger than 16 KB
The scoped connection reimplemented response conversion and only
handled byte arrays, stringifying the zero-copy DirectByteBuffer that
the native layer returns for any response over 16 KB. A large value
came back as the buffer's debug string, so callers silently received
corrupted data with no exception.

Route scope responses through the same decoder the normal client uses,
now extracted into a shared DirectBufferResolver, so large strings,
arrays, and maps decode identically on both paths. Add command() to
return the decoded object for aggregate replies alongside the existing
string-typed helpers.
